requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
alembic==1.15.2
Flask-Cors==4.0.0
SQLAlchemy~=2.0.40
PyJWT==2.10.1
numpy==2.2.5
